requirements.txt
django==4.0
channels==3.0.5
langchain-google-genai
google-generativeai
langchain
langchain-community
langchain-core
langchain-classic
faiss-cpu